+++ HetznerCloudApi/Core.cs
@@ -2,12 +2,15 @@
 using System.Net;
 using System.Threading.Tasks;
 using System;
+using HetznerCloudApi.Object.Universal;
+using Newtonsoft.Json.Linq;
+using Newtonsoft.Json;
 
 namespace HetznerCloudApi
 {
     public class Core
     {
-        public static long PerPage = 2;
+        public static long PerPage = 5;
 
         private const string ApiServer = "https://api.hetzner.cloud/v1";
 
@@ -23,17 +26,27 @@ namespace HetznerCloudApi
                 }
             }
 
-            // IsNullOrEmpty
+            // Response
             string json = await httpResponseMessage.Content.ReadAsStringAsync();
-            if (string.IsNullOrEmpty(json))
-            {
-                throw new Exception("there has been some error, the API has responded empty.");
-            }
 
             // No OK
             if (httpResponseMessage.StatusCode != HttpStatusCode.OK)
             {
-                throw new Exception("{error.Message}");
+                // Get Error
+                JObject result = JObject.Parse(json);
+                Error error = JsonConvert.DeserializeObject<Error>($"{result["error"]}") ?? new Error();
+
+                //Check error
+                if (error.Message.Contains("with ID") && error.Message.Contains("not found"))
+                {
+                    // The error is due to the resource not being found. Let's make it return empty instead of an error.
+                    json = "{}";
+                }
+                else
+                {
+                    // If it's a genuine error
+                    throw new Exception("{error.Message}");
+                }
             }
 
             return json;