+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.0.2 (2023-11-01)
+- ServerType; Complete
+- When searching for an object and if it doesn't exist, it used to throw an exception. It has been changed to return an empty object instead of an exception
+
 ## Version 0.0.1 (2023-10-31)
 - Image; Complete
 - Project initiation